Emotion Classification/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.1.0=pypi_0
astunparse=1.6.3=pypi_0
ca-certificates=2022.4.26=haa95532_0
cachetools=5.2.0=pypi_0
certifi=2022.5.18.1=py38haa95532_0
charset-normalizer=2.0.12=pypi_0
colorama=0.4.4=pypi_0
cycler=0.11.0=pypi_0
flatbuffers=1.12=pypi_0
fonttools=4.33.3=pypi_0
gast=0.4.0=pypi_0
google-auth=2.7.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.46.3=pypi_0
h5py=3.7.0=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.11.4=pypi_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.2=pypi_0
libclang=14.0.1=pypi_0
markdown=3.3.7=pypi_0
matplotlib=3.5.2=pypi_0
numpy=1.22.4=pypi_0
oauthlib=3.2.0=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1o=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
pillow=9.1.1=pypi_0
pip=21.2.2=py38haa95532_0
protobuf=3.19.4=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=3.0.9=pypi_0
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pypi_0
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
requests=2.28.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scipy=1.8.1=pypi_0
seaborn=0.11.2=pypi_0
setuptools=61.2.0=py38haa95532_0
six=1.16.0=pypi_0
sqlite=3.38.3=h2bbff1b_0
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.9.1=pypi_0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-io-gcs-filesystem=0.26.0=pypi_0
termcolor=1.1.0=pypi_0
torch=1.11.0=pypi_0
torchvision=0.12.0=pypi_0
tqdm=4.64.0=pypi_0
typing-extensions=4.2.0=pypi_0
urllib3=1.26.9=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
wrapt=1.14.1=pypi_0
zipp=3.8.0=pypi_0
